+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 2026-05-14
+### Fixed
+- Fixed assert-equal ansi check for different length strings.
+- Fixed reading characters '{' and '}'.
+### Added
+- Added the `bind` function for local variable binding in test suite setup functions.
+- Test filters can now included regular expressions by bracketing the
+  string with a '/' character. Some restrictions apply.
+- Added then `parse-time` function.
 
 ## [1.4.0] - 2026-04-21
 ### Fixed